+++ Plugin for WP sites/projects/rosenberger/theme/blocks/provision-callout/src/index.js
@@ -1,0 +1,5 @@
+import { registerBlockType } from '@wordpress/blocks';
+import metadata from '../block.json';
+import Edit from './edit';
+
+registerBlockType( metadata.name, { edit: Edit, save: () => null } );